Break layer render to MapData with question switch
